move buttons to their own components
