+++ frontend/src/components/GameShow/GameShow.css
@@ -71,12 +71,10 @@
     border-radius: 3px;
 }
 
-.owned-game-button,
 .wishlist-button {
 	cursor: pointer;
 }
 
-.owned-game-button:hover, 
 .wishlist-button:hover {
 	background-color: lightgray;
 }
