+++ frontend/src/components/WishlistButton/WishlistButton.css
@@ -1,0 +1,8 @@
+
+.wishlist-button {
+	cursor: pointer;
+}
+
+.wishlist-button:hover {
+	background-color: lightgray;
+}
